+++ src/components/molecules/Common/plugin/builtin/widgets/menu.stories.tsx
@@ -61,6 +61,11 @@ const property: Property = {
       buttonPosition: "bottomleft",
       buttonStyle: "texticon",
     },
+    {
+      id: "emptyurl",
+      buttonType: "link",
+      buttonTitle: "Empty URL",
+    },
   ],
   menu: [
     {
